% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.macie_operations.R
\name{associate_member_account}
\alias{associate_member_account}
\title{Associates a specified AWS account with Amazon Macie as a member account}
\usage{
associate_member_account(
  memberAccountId = "string"
)
}
\arguments{
\item{memberAccountId}{[required] The ID of the AWS account that you want to associate with Amazon Macie as a member account.}
}
\description{
Associates a specified AWS account with Amazon Macie as a member account.
}
\examples{

}
