% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.macie_operations.R
\name{disassociate_member_account}
\alias{disassociate_member_account}
\title{Removes the specified member account from Amazon Macie}
\usage{
disassociate_member_account(memberAccountId)
}
\arguments{
\item{memberAccountId}{[required] The ID of the member account that you want to remove from Amazon Macie.}
}
\description{
Removes the specified member account from Amazon Macie.
}
\section{Accepted Parameters}{
\preformatted{disassociate_member_account(
  memberAccountId = "string"
)
}
}

