% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.macie_operations.R
\name{disassociate_s3_resources}
\alias{disassociate_s3_resources}
\title{Removes specified S3 resources from being monitored by Amazon Macie}
\usage{
disassociate_s3_resources(associatedS3Resources, memberAccountId = NULL)
}
\arguments{
\item{associatedS3Resources}{The S3 resources (buckets or prefixes) that you want to remove from being monitored and classified by Amazon Macie.}

\item{memberAccountId}{The ID of the Amazon Macie member account whose resources you want to remove from being monitored by Amazon Macie.}
}
\description{
Removes specified S3 resources from being monitored by Amazon Macie. If memberAccountId isn't specified, the action removes specified S3 resources from Macie for the current master account. If memberAccountId is specified, the action removes specified S3 resources from Macie for the specified member account.
}
\examples{

}
