% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.macie_operations.R
\name{list_member_accounts}
\alias{list_member_accounts}
\title{Lists all Amazon Macie member accounts for the current Amazon Macie master account}
\usage{
list_member_accounts(
  nextToken = "string",
  maxResults = 123
)
}
\arguments{
\item{nextToken}{Use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListMemberAccounts action. Subsequent calls to the action fill nextToken in the request with the value of nextToken from the previous response to continue listing data.}

\item{maxResults}{Use this parameter to indicate the maximum number of items that you want in the response. The default value is 250.}
}
\description{
Lists all Amazon Macie member accounts for the current Amazon Macie master account.
}
\examples{

}
