% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.macie_operations.R
\name{list_s3_resources}
\alias{list_s3_resources}
\title{Lists all the S3 resources associated with Amazon Macie}
\usage{
list_s3_resources(memberAccountId = NULL, nextToken = NULL,
  maxResults = NULL)
}
\arguments{
\item{memberAccountId}{The Amazon Macie member account ID whose associated S3 resources you want to list.}

\item{nextToken}{Use this parameter when paginating results. Set its value to null on your first call to the ListS3Resources action. Subsequent calls to the action fill nextToken in the request with the value of nextToken from the previous response to continue listing data.}

\item{maxResults}{Use this parameter to indicate the maximum number of items that you want in the response. The default value is 250.}
}
\description{
Lists all the S3 resources associated with Amazon Macie. If memberAccountId isn't specified, the action lists the S3 resources associated with Amazon Macie for the current master account. If memberAccountId is specified, the action lists the S3 resources associated with Amazon Macie for the specified member account.
}
\examples{

}
